Rename event details form (#8)
